# public/swagger/swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": "backend api",
		"version": "1.0.0"
	},
	"paths": {
...
		"/password/email": {
			"post": {
				"summary": "Enviar email para resetear password",
				"description": "Enviar email para resetear password",
				"operationId": "sendResetLink",
				"consumes": [
					"application/x-www-form-urlencoded"
				],
				"parameters": [
					{
						"in": "formData",
						"name": "email",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Email enviado"
					},
					"400": {
						"description": "Falta email o email invalido"
					}
				}
			}
		},
		"/password/reset": {
			"post": {
				"summary": "Enviar email para resetear password",
				"description": "Enviar email para resetear password",
				"operationId": "passwordReset",
				"consumes": [
					"application/x-www-form-urlencoded"
				],
				"parameters": [
					{
						"in": "formData",
						"name": "email",
						"required": true,
						"type": "string"
					},
					{
						"in": "formData",
						"name": "token",
						"required": true,
						"type": "string"
					},
					{
						"in": "formData",
						"name": "password",
						"required": true,
						"type": "string"
					},
					{
						"in": "formData",
						"name": "password_confirmation",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Passwor modificado correctamente"
					},
					"400": {
						"description": "Falta token o token invalido"
					}
				}
			}
		},
...
	}
}